+++ packages/metadata-compatibility/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./isCompatible"
+export * from "./isStaticCompatible"
+export { mapLookupToTypedef, type TypedefNode } from "./typedef"