+++ packages/signers/pjs-signer/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Add API to sign raw-data and rename `sign` to `signTx`
+
 ## 0.2.1 - 2024-07-03
 
 ### Fixed